+++ src/pluckit/plugins/__init__.py
@@ -1,6 +1,6 @@
 from pluckit.plugins.base import Plugin, PluginRegistry
 from pluckit.plugins.history import Commit, History
-from pluckit.plugins.viewer import AstViewer, Rule, parse_viewer_query
+from pluckit.plugins.viewer import AstViewer, Rule, View, ViewBlock, parse_viewer_query
 
 __all__ = [
     "Plugin",
@@ -8,6 +8,8 @@ __all__ = [
     "AstViewer",
     "Rule",
     "parse_viewer_query",
+    "View",
+    "ViewBlock",
     "History",
     "Commit",
 ]